+++ app/appContainer.tsx
@@ -4,38 +4,12 @@ import css from "./AppContainer.module.scss";
 import Products from "./components/products";
 import Recipe from "./components/recipe";
 import recipeReducer from "./recipeReducer";
-import { Product } from "./globalTypes";
+import { ActionType, Product } from "./globalTypes";
 
 const initialRecipe: Product[] = [];
 
-export enum ActionType {
-  ADD = "add",
-  MODIFY = "modify",
-  DELETE = "delete",
-}
-
-type ActionAddProduct = {
-  type: ActionType.ADD;
-  newProduct: Product;
-};
-
-type ActionModifyAmount = {
-  type: ActionType.MODIFY;
-};
-
-type ActionDeleteProduct = {
-  type: ActionType.DELETE;
-};
-
-export type Action =
-  | ActionAddProduct
-  | ActionModifyAmount
-  | ActionDeleteProduct;
-
 const AppContainer = () => {
   const [recipe, dispatch] = useReducer(recipeReducer, initialRecipe);
-
-  console.log(recipe);
 
   const handleAddProduct = (product: Product) => {
     dispatch({
@@ -44,10 +18,17 @@ const AppContainer = () => {
     });
   };
 
+  const handleDeleteProduct = (productId: number) => {
+    dispatch({
+      type: ActionType.DELETE,
+      id: productId,
+    });
+  };
+
   return (
     <div className={css.app_container}>
       <Products addProduct={handleAddProduct} />
-      <Recipe />
+      <Recipe recipe={recipe} deleteProduct={handleDeleteProduct} />
     </div>
   );
 };
