+++ app/globalTypes.ts
@@ -10,3 +10,28 @@ export enum Unit {
   G = "g",
   ML = "ml",
 }
+
+export enum ActionType {
+  ADD = "add",
+  MODIFY = "modify",
+  DELETE = "delete",
+}
+
+type ActionAddProduct = {
+  type: ActionType.ADD;
+  newProduct: Product;
+};
+
+type ActionModifyAmount = {
+  type: ActionType.MODIFY;
+};
+
+type ActionDeleteProduct = {
+  type: ActionType.DELETE;
+  id: number;
+};
+
+export type Action =
+  | ActionAddProduct
+  | ActionModifyAmount
+  | ActionDeleteProduct;
